refactor: rename HeatmapViewer to CXR4HeatmapViewer
