+++ libs/insight-viewer/src/Viewer/CXR4HeatmapViewer/HeatmapViewer.types.ts
