+++ libs/insight-viewer/src/Viewer/CXR4HeatmapViewer/index.tsx
@@ -5,7 +5,7 @@ import { style } from './HeatmapViewer.styles'
 
 import type { HeatmapViewerProps } from './HeatmapViewer.types'
 
-export function HeatmapViewer(props: HeatmapViewerProps): ReactElement<HTMLCanvasElement> {
+export function CXR4HeatmapViewer(props: HeatmapViewerProps): ReactElement<HTMLCanvasElement> {
   const [canvasRef] = useHeatmapDrawing(props)
 
   return <canvas data-cy-name="heatmap-canvas" ref={canvasRef} style={style} />
