+++ libs/insight-viewer/src/Viewer/CXR4HeatmapViewer/useHeatmapDrawing.ts
@@ -4,8 +4,8 @@ import { useEffect, useMemo, useRef, RefObject } from 'react'
 import { HeatmapViewerProps } from './HeatmapViewer.types'
 import { useOverlayContext } from '../../contexts'
 
-import drawHeatmap from '../../utils/HeatmapViewer/drawHeatmap'
-import getHeatmapImageData from '../../utils/HeatmapViewer/getHeatmapImageData'
+import drawHeatmap from '../../utils/CXR4HeatmapViewer/drawHeatmap'
+import getHeatmapImageData from '../../utils/CXR4HeatmapViewer/getHeatmapImageData'
 
 function useHeatmapDrawing({ posMap, threshold }: HeatmapViewerProps): [RefObject<HTMLCanvasElement>] {
   const canvasRef = useRef<HTMLCanvasElement>(null)
