+++ libs/insight-viewer/src/index.ts
@@ -1,6 +1,6 @@
 export * from './const'
 export { InsightViewer as default } from './Viewer'
-export { HeatmapViewer } from './Viewer/HeatmapViewer'
+export { CXR4HeatmapViewer } from './Viewer/CXR4HeatmapViewer'
 export { LegacyHeatmapViewer } from './Viewer/LegacyHeatmapViewer'
 export { useMultipleImages } from './hooks/useMultipleImages'
 export { useInteraction } from './hooks/useInteraction'
