+++ libs/insight-viewer/src/utils/CXR4HeatmapViewer/clearHeatmap.ts
