+++ libs/insight-viewer/src/utils/CXR4HeatmapViewer/drawHeatmap.ts
