+++ libs/insight-viewer/src/utils/CXR4HeatmapViewer/getHeatmapImageData.ts
