+++ libs/insight-viewer/src/utils/CXR4HeatmapViewer/getRGBArray.ts
